Fixed bug with persite log likelihoods
persite_lnl arrays were filled using `m++` instead of the sites iterator.
In a few cases 'm' was not reinitialized to 0 before the loop and it could produce wrong site likelihoods and occasionally a segmentation fault.
Array indices were set to the site iterator in order to avoid further errors.
This fix also fixes occasional errors in the testing framework.